Renamed ActorGroupManager.h/.cpp to SpatialActorGroupManager.h/.cpp. Renamed SnapshotManager.h/.cpp to SpatialSnapshotManager.h/.cpp
