+++ SpatialGDK/Source/SpatialGDK/Private/Interop/SpatialSnapshotManager.cpp
@@ -1,6 +1,6 @@
 // Copyright (c) Improbable Worlds Ltd, All Rights Reserved
 
-#include "Interop/SnapshotManager.h"
+#include "Interop/SpatialSnapshotManager.h"
 
 #include "Interop/Connection/SpatialWorkerConnection.h"
 #include "Interop/GlobalStateManager.h"
