+++ SpatialGDK/Source/SpatialGDK/Private/Utils/SpatialActorGroupManager.cpp
@@ -1,6 +1,6 @@
 // Copyright (c) Improbable Worlds Ltd, All Rights Reserved
 
-#include "Utils/ActorGroupManager.h"
+#include "Utils/SpatialActorGroupManager.h"
 #include "SpatialGDKSettings.h"
 
 void SpatialActorGroupManager::Init()
