+++ SpatialGDK/Source/SpatialGDK/Public/Interop/SpatialSnapshotManager.h
