+++ SpatialGDK/Source/SpatialGDK/Public/SpatialGDKSettings.h
@@ -5,7 +5,7 @@
 #include "CoreMinimal.h"
 #include "Engine/EngineTypes.h"
 #include "Misc/Paths.h"
-#include "Utils/ActorGroupManager.h"
+#include "Utils/SpatialActorGroupManager.h"
 #include "Utils/SpatialDebugger.h"
 
 #include "SpatialGDKSettings.generated.h"
